bot/requirements.txt
aio_pika
aiogram
python-dotenv
redis
asyncio
